agent/requirements.txt
livekit-agents[silero,turn-detector,openai]~=1.3
livekit-plugins-noise-cancellation~=0.2
python-dotenv==1.2.1
piper-tts==1.3.0